requirements.txt
transformers==4.39.3
accelerate==0.28.0
sentencepiece==0.2.0
torchmetrics==1.3.2
datasets==2.18.0
einops==0.8.0
einops_exts==0.0.4
evaluate==0.4.3
rouge_score==0.1.2
peft==0.11.1